Api: final settings

+++ src/main/java/dev/project/servicesweb/repositories/CompteRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CompteRepository extends JpaRepository<Compte, Integer> {
+public interface CompteRepository extends JpaRepository<Compte, Long> {
 
 }
